+++ ShingoEvents/EventsTableViewController.swift
@@ -14,8 +14,10 @@ class EventsTableViewController: UITableViewController {
     
     // MARK: - Properties
     var events: [SIEvent]!
-
+    
     var gradientBackgroundView = UIView()
+    
+    var advertIsDonePresenting: Bool = false
     
     override func viewWillAppear(_ animated: Bool) {
         super.viewWillAppear(animated)
@@ -54,7 +56,7 @@ class EventsTableViewController: UITableViewController {
     }
 }
 
-extension EventsTableViewController: SICellDelegate {
+extension EventsTableViewController: SICellDelegate, SplashScreenViewDelegate {
 
     // MARK: - Table view data source
     
@@ -89,22 +91,53 @@ extension EventsTableViewController: SICellDelegate {
         
         let event = events[(indexPath as NSIndexPath).row]
         
-        if event.didLoadEventData {
-            self.performSegue(withIdentifier: "EventMenu", sender: event)
-            
+        self.performSegue(withIdentifier: "EventMenu", sender: event)
+    }
+    
+    override func performSegue(withIdentifier identifier: String, sender: Any?) {
+        
+        guard let event = sender as? SIEvent else {
+            return
+        }
+        
+        let splashScreen = SplashScreenView(viewController: self, identifier: identifier, event: event)
+        event.didDisplaySponsorAd = true
+        if event.didLoadEventData && event.didDisplaySponsorAd {
+            super.performSegue(withIdentifier: identifier, sender: event)
+        } else if event.didLoadEventData {
+            present(splashScreen, animated: true) {
+                self.onPresentSplashScreenComplete(identifier: identifier, event: event)
+            }
         } else {
-            let activityView = ActivityViewController(message: "Grabbing Event Data...")
-            present(activityView, animated: false, completion: {
+            present(splashScreen, animated: true) {
                 event.requestEvent() {
-                    self.dismiss(animated: true, completion: {
-                        if event.didLoadEventData {
-                            self.performSegue(withIdentifier: "EventMenu", sender: self.events[(indexPath as NSIndexPath).row])
-                        } else {
-                            self.displayBadRequestNotification()
-                        }
-                    })
+                    print("STOP Request Event")
+                    if !event.didLoadEventData {
+                        self.displayBadRequestNotification()
+                    }
+                    self.onPresentSplashScreenComplete(identifier: identifier, event: event)
                 }
-            })
+            }
+        }
+    }
+    
+    /**
+     This function allows the view to dismiss a `SplashScreenView` and perform a segue at the correct time.
+     The view will dismiss a `SplashScreenView` and let the view controller perform a segue when:
+     
+     a) An event has already loaded its data, but still needs to display a sponsor ad, or
+     
+     b) A sponsor ad has been displayed for a predetermined length of time but the event is still loading its data
+     
+     - parameter identifier: The identfier of the viewController to be segued to.
+     - parameter event: The event object to be displayed on the following viewController.
+     */
+    func onPresentSplashScreenComplete(identifier: String, event: SIEvent) {
+        if event.didDisplaySponsorAd && event.didLoadEventData {
+            dismiss(animated: false) {
+                super.performSegue(withIdentifier: identifier, sender: event)
+            }
+            event.didDisplaySponsorAd = false
         }
     }
