+++ ShingoEvents/SplashScreenView.swift
@@ -9,22 +9,43 @@
 import Foundation
 import UIKit
 
+protocol SplashScreenViewDelegate {
+    func onPresentSplashScreenComplete(identifier: String, event: SIEvent)
+}
+
 class SplashScreenView: UIViewController {
     
-    let minPresentationLength: TimeInterval = 3 * 1000
+    static let defaultPresentationLength: TimeInterval = TimeInterval(exactly: 3.0)!
     
-    var canDismissSplashScreen = false
+    var presentedAt: Date?
     
-    var imageView: UIImageView! {
-        didSet {
-            imageView.contentMode = UIViewContentMode.scaleAspectFit
-            imageView.backgroundColor = .black
-        }
-    }
+    var timer: Timer!
+    
+    var imageView: UIImageView!
+    
+    var delegate: SplashScreenViewDelegate?
+    
+    var identifier: String!
+    
+    var event: SIEvent!
     
     var didUpdateViewConstrains = false
     
-    public override init(nibName nibNameOrNil: String?, bundle nibBundleOrNil: Bundle?) {
+    /**
+     - parameter splashScreenImage: The advertisement image to be displayed on the screen
+     - parameter delegate: The view controller currently presenting the splash screen.
+     */
+    convenience init(viewController parent: SplashScreenViewDelegate, identifier: String, event: SIEvent) {
+        self.init(nibName: nil, bundle: nil)
+        self.delegate = parent
+        self.imageView = UIImageView()
+        self.imageView.contentMode = UIViewContentMode.scaleAspectFit
+        self.imageView.backgroundColor = .black
+        self.identifier = identifier
+        self.event = event
+    }
+    
+    fileprivate override init(nibName nibNameOrNil: String?, bundle nibBundleOrNil: Bundle?) {
         super.init(nibName: nibNameOrNil, bundle: nibBundleOrNil)
         
         // default modal transition style should be `crossDissolve`.
@@ -32,11 +53,6 @@ class SplashScreenView: UIViewController {
         
         // default modal presentation style should be `overCurrentContext`.
         modalPresentationStyle = .overCurrentContext
-    }
-    
-    convenience init(splashScreenImage: UIImage) {
-        self.init(nibName: nil, bundle: nil)
-        self.imageView = UIImageView(image: splashScreenImage)
     }
     
     required init?(coder aDecoder: NSCoder) {
@@ -47,78 +63,59 @@ class SplashScreenView: UIViewController {
         super.viewDidLoad()
         view.addSubview(imageView)
         imageView.contentMode = .scaleAspectFit
-        imageView.backgroundColor = .black
         updateViewConstraints()
+    }
+    
+    func setSponsorAd(for event: SIEvent) {
+        
+        guard let splashAd = event.getSplashAd() else {
+            imageView.image = #imageLiteral(resourceName: "Shingo Icon Fullscreen")
+            return
+        }
+        
+        if let image = splashAd.image {
+            splashAd.presentedAt = Date()
+            imageView.image = image
+        } else {
+            splashAd.makeImageRequest()
+        }
+
+    }
+    
+    override func viewWillAppear(_ animated: Bool) {
+        super.viewWillAppear(animated)
+        setSponsorAd(for: event)
+    }
+    
+    override func viewDidAppear(_ animated: Bool) {
+        super.viewDidAppear(animated)
+        if #available(iOS 10.0, *) {
+            Timer.scheduledTimer(withTimeInterval: SplashScreenView.defaultPresentationLength, repeats: true) { (_) in
+                self.onMinPresentationTimerComplete()
+            }
+        } else {
+            Timer.scheduledTimer(timeInterval: SplashScreenView.defaultPresentationLength,
+                                 target: self,
+                                 selector: #selector(self.onMinPresentationTimerComplete),
+                                 userInfo: nil,
+                                 repeats: false)
+        }
+        
     }
     
     override func updateViewConstraints() {
         if !didUpdateViewConstrains {
-            
             imageView.autoPinEdgesToSuperviewEdges()
-            
             didUpdateViewConstrains = true
         }
         super.updateViewConstraints()
     }
     
-    func presentSplashScreen(parentViewController parent: UIViewController, completion: @escaping ((Void) -> Void)) {
-        parent.present(self, animated: true) {
-            self.presentationTimer = Timer.scheduledTimer(
-                timeInterval: 3.0,
-                target: self,
-                selector: #selector(self.onMinPresentationLength),
-                userInfo: nil,
-                repeats: true
-            )
-            
-            self.presentationTimer.fire()
-            
-            completion()
-            
+    func onMinPresentationTimerComplete() {
+        event.didDisplaySponsorAd = true
+        if let delegate = self.delegate {
+            delegate.onPresentSplashScreenComplete(identifier: identifier, event: event)
         }
-    }
-    
-    var parentVC: UIViewController!
-    var onCompletion: ((Void) -> Void)!
-    var waitTimer: Timer!
-    var presentationTimer: Timer!
-    var timerFireCount = 0
-    
-    
-    func dismissSplashScreen(parentViewController parent: UIViewController, completion: @escaping ((Void) -> Void)) {
-        
-        if parentVC == nil {
-            parentVC = parent
-        }
-        
-        if onCompletion == nil {
-            onCompletion = completion
-        }
-        
-        if canDismissSplashScreen {
-           parentVC.dismiss(animated: true, completion: onCompletion)
-        } else {
-            waitTimer = Timer.scheduledTimer(timeInterval: 0.5, target: self, selector: #selector(self.waitForMinPresentationTime), userInfo: nil, repeats: true)
-        }
-    }
-    
-    func waitForMinPresentationTime() {
-        if canDismissSplashScreen {
-            waitTimer.invalidate()
-            parentVC.dismiss(animated: false, completion: onCompletion)
-        }
-    }
-    
-    func onMinPresentationLength() {
-        
-        if timerFireCount > 0 {
-            print("\(3 * timerFireCount) Seconds have passed!")
-            canDismissSplashScreen = true
-            presentationTimer.invalidate()
-        } else {
-            timerFireCount += 1
-        }
-        
     }
     
 }
